# Use multi-core optimized pgzip package in tarball compressor