Some early benchmarks (when using these changes in the bosh-cli):
```
time bosh create-release --dir ~/workspace/carvel-on-bosh-demo/release --force --tarball=/tmp/slow.tgz
# 5.69s user 0.66s system 96% cpu 6.569 total
time ./out/bosh create-release --dir ~/workspace/carvel-on-bosh-demo/release --force --tarball=/tmp/fast.tgz
# 1.11s user 0.53s system 105% cpu 1.556 total
```

I did make this a breaking change to the interface because we no longer need the CmdRunner in the TarballCompressor.