@rkoster 2024-10-17
> Is the `pgzip` library actively maintained? I was checking the activity in the repository and there were not much since two years. It could be that there was no need for any updates because it only adds parallel execution.

The maintainer is still active on issues (the most recent one is from February): https://github.com/klauspost/pgzip/issues/56#issuecomment-1969471820

So to me that suggests the project is feature-complete. It also has a healthy number of github stars.